% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_hal_augment.R
\name{fit_hal_augment}
\alias{fit_hal_augment}
\title{Fit HAL using Data Augmentation}
\usage{
fit_hal_augment(X, Y, col_idx, noise, copies, basis_list, ...)
}
\arguments{
\item{X}{An input `matrix` with dimension number of observations -by-
number of covariates.}

\item{Y}{A `numeric` vector of observations of the outcome variable.}

\item{col_idx}{A `numeric` vector of column indices to apply noise.}

\item{noise}{A `numeric` value for the standard deviation of the Gaussian
noise.}

\item{copies}{A `numeric` value for the number of augmented data sets.}

\item{basis_list}{A `list` of basis functions generated from `X`.}

\item{...}{Additional arguments to be passed to `fit_hal()`.}
}
\value{
A `list` with the following components:
\item{X}{An augmented `matrix` of the original design matrix `X`.}
\item{Y}{An augmented `numeric` vector of the outcome variable `Y`. Note
that this simply contains the original `Y` repeated `copies + 1` times.}
\item{fit}{An object returned from the function `fit_hal()`.}
}
\description{
A wrapper function for `fit_hal()`. Generate augmented data
sets by applying Gaussian noise to the original design matrix `X`.
}
