% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_num_basis}
\alias{get_num_basis}
\title{Function to get the number of HAL basis functions of each degree
of a given basis list}
\usage{
get_num_basis(basis_list)
}
\description{
Function to get the number of HAL basis functions of each degree
of a given basis list
}
