% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_smallest_eigen_hal9001}
\alias{get_smallest_eigen_hal9001}
\title{Function to get the smallest eigenvalue of X^TX, where X is formed by
selected bases from hal9001 fit}
\usage{
get_smallest_eigen_hal9001(fit, X, add_main)
}
\description{
Function to get the smallest eigenvalue of X^TX, where X is formed by
selected bases from hal9001 fit
}
